Code improvements (#234)
* Update to use best practice (more performant) ways for creating/adding to collections.

* A number of changes including removing unneeded explicit boxing.unboxing, simplifying for loops (and in many cases being able to get rid of a now unused size variable), changing to use text blocks (less concatenation), changing some if/else chains to switch statements, using try-with-resources instead of finally blocks, and other improvements.

* A number of code simplifications (removing redundant casts, unneeded toStrings, redundant checks, simplifying a number of string operations, consolidated duplicated switch cases. The main non-simplification is adding final where possible, which allows for better optimizations (in the compiler)

* Change the new public static final arrays to use ImmutableList (ErrorProne warning)

More simplifications including skipping unneeded variable declarations (like those that are then returned the next line), changed a couple internal classes to records.

* Finish the JdkObsolete refactor (ErrorProne errrors). This is mainly a performance focused change.

* Set charset used in locations that were defaulting to platform standard. This should help ensure behavior stays consistent across platforms. (ErrorProne warning)

More code simplification as well, including consolidation of duplicate code across if/else branches (and other similar simplifications).

* Fix a data license that changed on an external server.

+++ WEB-INF/classes/dods/dap/DInt16.java
@@ -149,6 +149,6 @@ public class DInt16 extends BaseType implements ClientIO {
     // Write this value out as a 32bit int, since the smallest thing
     // we write to the network is 32bits. The cast automagically
     // pads it out with zero's
-    sink.writeInt((int) val);
+    sink.writeInt(val);
   }
 }
